backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "channels>=4.3.2",
    "daphne>=4.2.1",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "langchain[google-genai]>=1.2.1",
    "langchain-core>=1.2.6",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "load-dotenv>=0.1.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pandas>=2.3.3",
    "langsmith>=0.6.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = ["F401"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []


